+++ src/onegov.gazette/onegov/gazette/tests/test_forms.py
@@ -341,8 +341,12 @@ def test_user_form(session):
     # Test apply / update
     users = UserCollection(session)
     user = users.add(
-        username='a@a.ai', realname='User', role='editor', password='pwd'
+        username='a@a.ai',
+        realname='User',
+        role='editor',
+        password='pwd',
     )
+    user.phone_number = '+41415554433'
 
     groups = UserGroupCollection(session)
     group = groups.add(name='Group A')
@@ -354,6 +358,7 @@ def test_user_form(session):
     assert form.role.data == 'editor'
     assert form.name.data == 'User'
     assert form.group.data == ''
+    assert form.phone_number.data == '+41415554433'
 
     user.group = group
     session.flush()
@@ -365,12 +370,14 @@ def test_user_form(session):
     form.role.data = 'publisher'
     form.name.data = 'Publisher'
     form.group.data = ''
+    form.phone_number.data = '0415554434'
 
     form.update_model(user)
     assert user.username == 'b@b.bi'
     assert user.role == 'publisher'
     assert user.realname == 'Publisher'
     assert user.group_id is None
+    assert user.phone_number == '+41415554434'
 
     session.flush()
     session.refresh(user)
@@ -446,9 +453,14 @@ def test_user_form_on_request(session):
 
 
 def test_notice_form(session, categories, organizations, issues):
-    # Test apply / update
-    form = NoticeForm()
-    form.request = DummyRequest(session)
+    users = UserCollection(session)
+    user = users.add(
+        username='a@a.ai',
+        realname='User',
+        role='editor',
+        password='pwd',
+    )
+    user.phone_number = '+41415554433'
 
     notice = GazetteNotice(
         title='Title',
@@ -460,9 +472,11 @@ def test_notice_form(session, categories, organizations, issues):
     notice.organization_id = '200'
     notice.category_id = '13'
     notice.issues = ['2017-43']
+    notice.user = user
 
-    notice.author_date is None
-    notice.author_name is None
+    # Test apply / update
+    form = NoticeForm()
+    form.request = DummyRequest(session)
 
     form.apply_model(notice)
     assert form.title.data == 'Title'
@@ -478,6 +492,7 @@ def test_notice_form(session, categories, organizations, issues):
         datetime(2018, 1, 1), 'UTC'
     )
     assert form.issues.data == ['2017-43']
+    assert form.phone_number.data == '+41415554433'
 
     form.title.data = 'Notice'
     form.organization.data = '300'
@@ -490,6 +505,7 @@ def test_notice_form(session, categories, organizations, issues):
     form.author_name.data = 'Bureau of Public Affairs'
     form.author_date.data = standardize_date(datetime(2019, 1, 1), 'UTC')
     form.issues.data = ['2017-44']
+    form.phone_number.data = '796662211'
 
     form.update_model(notice)
     assert notice.title == 'Notice'
@@ -504,6 +520,7 @@ def test_notice_form(session, categories, organizations, issues):
     assert notice.author_date == standardize_date(datetime(2019, 1, 1), 'UTC')
     assert notice.issues == {'2017-44': None}
     assert notice.first_issue == standardize_date(datetime(2017, 11, 3), 'UTC')
+    assert user.phone_number == '+41796662211'
 
     # Test validation
     form = NoticeForm()
@@ -606,14 +623,24 @@ def test_notice_form(session, categories, organizations, issues):
 
 
 def test_unrestricted_notice_form(session, categories, organizations, issues):
-    # Test apply / update
-    form = UnrestrictedNoticeForm()
-    form.request = DummyRequest(session)
+    users = UserCollection(session)
+    user = users.add(
+        username='a@a.ai',
+        realname='User',
+        role='editor',
+        password='pwd',
+    )
+    user.phone_number = '+41415554433'
 
     notice = GazetteNotice(title='Title', text='A <b>text</b>.')
     notice.organization_id = '200'
     notice.category_id = '13'
     notice.issues = ['2017-43']
+    notice.user = user
+
+    # Test apply / update
+    form = UnrestrictedNoticeForm()
+    form.request = DummyRequest(session)
 
     form.apply_model(notice)
     assert form.title.data == 'Title'
@@ -621,12 +648,14 @@ def test_unrestricted_notice_form(session, categories, organizations, issues):
     assert form.category.data == '13'
     assert form.text.data == 'A <b>text</b>.'
     assert form.issues.data == ['2017-43']
+    assert form.phone_number.data == '+41415554433'
 
     form.title.data = 'Notice'
     form.organization.data = '300'
     form.category.data = '11'
     form.text.data = 'A <b>notice</b>.'
     form.issues.data = ['2017-44']
+    form.phone_number.data = '796662211'
 
     form.update_model(notice)
     assert notice.title == 'Notice'
@@ -635,6 +664,7 @@ def test_unrestricted_notice_form(session, categories, organizations, issues):
     assert notice.text == 'A <b>notice</b>.'
     assert notice.issues == {'2017-44': None}
     assert notice.first_issue == standardize_date(datetime(2017, 11, 3), 'UTC')
+    assert user.phone_number == '+41796662211'
 
     notice.state = 'published'
     form.issues.data = ['2017-45']